removed layout.tsx and made page.tsx into a server component
